Refactor message handling structure and update import paths
- Introduced a new `message` package within the `utils` directory to encapsulate message-related functionalities, including `Message`, `MessageType`, `MessageManager`, and `status_manager`.
- Updated import paths across various modules to reflect the new structure, ensuring consistent access to message functionalities.
- Enhanced the Python import helper with updated paths for improved code completion related to message handling.
