+++ utils/global_vars.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
     from utils.settings import Settings
     from ui_files.ui_main_window import Ui_Form
     from ui_files.BlinkingLabel import BlinkingLabel
-    from utils.message_manager import MessageManager
+    from .message.message_manager import MessageManager
     from utils.robot.robot_status_monitor import RobotStatus
 
 from .logging_config import logger
