+++ utils/server/UR20_Server_functions.py
@@ -5,7 +5,7 @@ from PySide6.QtGui import QPixmap
 import utils
 from utils import global_vars
 from datetime import datetime
-from utils.status_manager import update_status_label
+from ..message.status_manager import update_status_label
 from typing import Literal, cast, Union
 from PySide6.QtCore import Qt, QObject, Signal, QTimer
 import logging
