+++ utils/ui/ui_helpers.py
@@ -10,12 +10,12 @@ from typing import Optional
 
 from utils import global_vars
 from ui_files.BlinkingLabel import BlinkingLabel
-from utils.message import MessageType
-from utils.message_manager import MessageManager
+from ..message.message import MessageType
+from ..message.message_manager import MessageManager
 from ui_files.PasswordDialog import PasswordEntryDialog
 from utils.robot.robot_control import load_rob_files
 from utils.usb_key_check import check_any_usb_for_key
-from utils.status_manager import update_status_label
+from ..message.status_manager import update_status_label
 from utils.robot.robot_control import load_wordlist
 from utils.ui.notification_popup import check_zwischenlage_status
